Introduce ThreadCleaner and use it in PooledByteBufAllocator to cleanup ThreadLocal caches.
Motivation:

At the moment we use ThreadDeathWatcher to check if we need to release buffers back to the pool when thread local caches are collected. In the past ThreadDeathWatcher was used which will need to "wakeup" every second to check if the registered Threads are still alive.

Modifications:

- Introduce ThreadCleaner and use it to release memory back to pool
- Deprecate ThreadDeathWatcher
- Add unit tests.
- Not hold a strong reference to the Thread in Recycler.Stack to allow it to be collected (IMHO this was a bug before as well to not do this.

Result:

Cheaper and better way to cleanup Thread resources.